Working automation for Upcoming Games Scraper + calculating predictions
