+++ web_app/data-automation/calculate-predictions.js
@@ -1,7 +1,5 @@
 const mysql = require('mysql');
 const connection = require("../database/connection.json");
-const upcomingGamesParser = require("../scrapers/player/upcoming-games-store.js");
-//import { weatherForecastData } from "../scrapers/weather/weather-forecast-script.js";
 const moment = require("moment");
 const http = require("http");
 const querystring = require("querystring");
@@ -14,50 +12,27 @@ const con = mysql.createConnection({
 	database: connection.database
 });
 
-//updateUpcomingGames();
-calculatePredictions();
+loop();
 
-/*async function updateUpcomingGames(){
+async function loop(){
 
-    //Deletes all current game predictions from DB
-    await con.query('DELETE FROM predictedgames', (err, rows, fields) => {
-        if (err) throw err;
-    });
+    //Automatically Recalculates Predictions Every 24 Hours
+    setInterval(calculatePredictions, 1000 * 60 * 60 * 24);
+}
 
-    //Updates Games from Upcoming to final status after the start time date passes
-    con.query('SELECT * FROM games WHERE status = "Upcoming"', (err, rows, fields) => {
-        if (err) throw err;
-
-        let games = rows;
-        let date = new Date().toJSON();
-        let dateTime = date.substring(0, 19);
-        for (let i = 0; i < games.length; i++){
-            s = moment(games[i].startTime).toISOString(true);
-            var startTime = s.substring(0, 19);
-            if (moment(startTime).isBefore(dateTime) == true){
-                con.query('UPDATE games SET status = "Final" WHERE gameID = ?',[games[i].gameID], (err, rows, fields) => {
-                    if (err) throw err;
-                });
-            }
-        }
-    }) 
-
-    await upcomingGamesParser.storeGamesOnPage;
-    //await calculatePredictions();
-    //weatherForecastData.gameDataRetrieval;
-    return;
-} */
-
-//Calculates Predictions for Current Upcoming Games
-//Stores Predicted Information into predictedgames table
 async function calculatePredictions() {
 
+    //Delete All current predictions stored in DB
+    await con.query('DELETE FROM predictedgames', (err, rows, fields) => {
+        if (err) throw err;
+    })
+
     //Get all upcoming games from DB
-    con.query('SELECT * FROM games WHERE status = "Upcoming"', (err, rows, fields) => {
+    await con.query('SELECT * FROM games WHERE status = "Upcoming"', (err, rows, fields) => {
 		if (err) throw err;
 
         let games = rows;
-        for (let i = 0; i < games.length; i++){
+        for (let i = 4; i < games.length; i++){
 
             //Argument sent to Prediction Algorithm
             //Will need to be changed when ML is updated
@@ -71,12 +46,12 @@ async function calculatePredictions() {
             };
 
             //Send request to Prediction API
-            const url_args = querystring.stringify(url_object);
+            const url_args =  querystring.stringify(url_object);
             let prediction = http.get("http://localhost:50300/api/v1/predict/teams/?" + url_args, response => {
                 let data = "";
                 response.on("data", chunk => {data += chunk});
                 response.on("end", () => {
-                    //console.log(data);
+                
                     let winner = JSON.parse(data).winner;
                     console.log(winner);
                     
@@ -108,7 +83,6 @@ async function calculatePredictions() {
                 });
             });
         }
-    
     });
 
     return;
