+++ web_app/scrapers/player/upcoming-games-store.js
@@ -1,5 +1,6 @@
 const internet = require('https')
 const mysql = require('mysql');
+const moment = require("moment");
 const connection = require("../../database/connection.json");
 const baseUrl = 'https://www.backend.audlstats.com/web-api/games?current'
 
@@ -13,10 +14,46 @@ const con = mysql.createConnection({
 	database: connection.database
 });
 
+loop();
+
+async function loop(){
+    //Gather new Upcoming Games from the AUDL every 24 hours
+    setInterval(storeGamesOnPage, 5000);
+}
+
 
 //Collect Upcoming Games Data for the next week from AUDL and store in DB
 const savePageJson = async function (res) {
     let body = "";
+
+    //Deletes Any remaining predictedgames from DB
+    await con.query('DELETE FROM predictedgames', (err, rows, fields) => {
+        if (err) throw err;
+    });
+
+    //Deletes Any leftover upcoming games from DB
+    await con.query('DELETE FROM games WHERE status = "Upcoming"', (err, rows, fields) => {
+        if (err) throw err;
+    });
+
+    //Updates Games from Upcoming to final status after the start time date passes
+    /*await con.query('SELECT * FROM games WHERE status = "Upcoming"', (err, rows, fields) => {
+        if (err) throw err;
+
+        let games = rows;
+        let date = new Date().toJSON();
+        let dateTime = date.substring(0, 19);
+        for (let i = 0; i < games.length; i++){
+            s = moment(games[i].startTime).toISOString(true);
+            var startTime = s.substring(0, 19);
+            if (moment(startTime).isBefore(dateTime) == true){
+                con.query('UPDATE games SET status = "Final" WHERE gameID = ?',[games[i].gameID], (err, rows, fields) => {
+                    if (err) throw err;
+                });
+            }
+        }
+    })  */
+
     try {
         //Wait until the information received from the https request is the actual raw data, then store it in a temporary variable
         await res.on("data", (chunk) => {
@@ -52,8 +89,6 @@ const savePageJson = async function (res) {
                         console.log(result);
                     });
             }
-            //Close DB Connection
-            con.end();
         });
     }catch (error){}
     return;
@@ -64,5 +99,3 @@ async function storeGamesOnPage (){
     await internet.get(baseUrl, savePageJson).on("error", (error) => {});
     return;
 }
-
-storeGamesOnPage();
